+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog
 
+## develop (unreleased)
+
+
+## 0.5.1 (4/15/2014)
+
+### Bugs fixed
+* Set hello message type = 0.
+
+
+## 0.5.0 (4/14/2014)
+
+### New features
+* Added new class `Pio::Hello`.
+
+### Misc
+* Added new rake task `cucumber`.
+* Added new rake task `dump_pcap`.
+
+
 ## 0.4.0 (3/31/2014)
 
 ### New features